Minor tweaks

+++ src/am_map.c
@@ -1394,7 +1394,7 @@ static void AM_DrawFline(int x0, int y0, int x1, int y1, byte *color,
 
                 while (x0 != x1)
                 {
-                    const int   mask = ~(error >> 31);
+                    const int   mask = ~((int64_t)error >> 31);
 
                     putdot((x0 += sx), (y0 += (sy & mask)), color);
                     error += dy - (dx & mask);
@@ -1410,7 +1410,7 @@ static void AM_DrawFline(int x0, int y0, int x1, int y1, byte *color,
 
                 while (y0 != y1)
                 {
-                    const int   mask = ~(error >> 31);
+                    const int   mask = ~((int64_t)error >> 31);
 
                     putdot((x0 += (sx & mask)), (y0 += sy), color);
                     error += dx - (dy & mask);
